+++ runtime/index.js
@@ -19,4 +19,5 @@ export * from "./nav.js";
 export * from "./lsnav.js";
 export * from "./diagramzoom.js";
 export * from "./tablescroll.js";
+export * from "./minimap.js";
 export * from "./tooltip.js";